fix: fixed a bug in infinite tictactoe
The bug was quite hard to detect, but i found it during
one of my game tests wherein logic_grid was not updating
even after me clicking on the cell. The game was making
corresponding sound. The reason was that i forgot to
reset the parameters for infinite tictactoe.
Also changed the class code to reduce lines.